+++ pepper-sync/src/sync/state.rs
@@ -824,7 +824,7 @@ where
     W: SyncWallet + SyncBlocks,
 {
     let sync_state = wallet.get_sync_state().map_err(SyncError::WalletError)?;
-    let start_height = sync_state
+    let birthday = sync_state
         .wallet_birthday()
         .expect("scan ranges must be non-empty");
     let fully_scanned_height = sync_state
@@ -834,7 +834,7 @@ where
     let (previously_scanned_sapling_outputs, previously_scanned_orchard_outputs) =
         calculate_scanned_outputs(wallet).map_err(SyncError::WalletError)?;
     let (birthday_sapling_initial_tree_size, birthday_orchard_initial_tree_size) =
-        if let Ok(block) = wallet.get_wallet_block(start_height) {
+        if let Ok(block) = wallet.get_wallet_block(birthday) {
             (
                 block.tree_bounds.sapling_initial_tree_size,
                 block.tree_bounds.orchard_initial_tree_size,
@@ -844,7 +844,7 @@ where
                 consensus_parameters,
                 fetch_request_sender.clone(),
                 wallet,
-                start_height - 1,
+                birthday - 1,
             )
             .await?
         };