+++ zingo-status/CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 `confirmation_status::ConfirmationStatus`:
 - added `Failed` variant
-- re-ordered variants to match serialization and the actual status flow calculate -> transmit -> mempool -> confirmed
+- new serialized version 1 with new variant
 
 ### Removed
 `confirmation_status::ConfirmationStatus::from_blockheight_and_pending_bool`